(#1) Install and use eslint via choco-theme
This installs eslint and applies it to files. A few files have been
updated due to the eslint config rules. This also adds choco-theme as a
devDependency so that eslint packages can be consumed by it, and so that
choco-theme functions or additional packages can be used inside of tests.

By running the choco-theme command, the eslint.config.mjs file will be
coppied over, ready for use.