(#5) Create a global theme toggle test
This change creates a test that will work on all Chocolatey websites,
except Chocolatey Central Management, that will assert that the theme
can be toggled from light to dark and vice versa. The locator to find
the theme toggle button has been updated so that it finds the toggle
button that is currently visible. Since some websites have different
buttons based on viewport, this will work no matter what size the
screen is at.